Migration fixes past 2014 (#162)

+++ database/migrations/2015_05_19_000002_hws_ssl_hostnames_drop_hostname_id.php
@@ -13,10 +13,12 @@ class HwsSslHostnamesDropHostnameId extends Migration
      */
     public function up()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->table('ssl_hostnames', function (Blueprint $table) {
-            // domain relation
-            $table->dropColumn('hostname_id');
-        });
+        if(Schema::connection(DatabaseConnection::systemConnectionName())->hasColumn('ssl_hostnames','hostname_id')){
+            Schema::connection(DatabaseConnection::systemConnectionName())->table('ssl_hostnames', function (Blueprint $table) {
+                // domain relation
+                $table->dropColumn('hostname_id');
+            });
+        }
     }
 
     /**
